Write Win32::Process::Info->variant_support_status($variant), which returns false if the variant is supported, and a message if not. Re-implement in terms of this, rather than _check_variant($variant), which throws an exception. The rest of the public interface did not change -- that is, public code which threw exceptions when a variant was unsupported still throws exceptions.
Also made the support check for the WMI variant like the check in
t/basic.t, which was more comprehensive. The hope is to recode t/basic.t
in terms of variant_support_status(), so that we are consistent about
what is supported and what is not.

Got rid of some dead code while I was about it.


git-svn-id: file:///Users/tom/Code/svn/repos/Win32-Process-Info/trunk@1748 f6fd3633-1913-47c2-bad4-372d75c948ec